requirements.txt
streamlit>=1.35
openai>=1.15
pandas
sqlalchemy
python-dotenv
matplotlib
